# backend/application/src/main/resources/api/v1/specs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Onboarding Platform API
  version: 1.0.0
  description: Backend API for AI Onboarding & Training platform
servers:
  - url: /api/v1
    description: V1 base path (documentation only)
...
paths:
...
  /api/v1/users/me:
    get:
      summary: Get my profile
      description: Returns the signed-in user's profile including role and avatar settings.
      tags: [ users ]
      operationId: getMyProfile
      security: [ { bearerAuth: [ ] } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UserProfileV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
    patch:
      summary: Update my profile
      description: Updates the signed-in user's display name, position, and avatar settings.
      tags: [ users ]
      operationId: updateMyProfile
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        description: Request payload for the updateMyProfile.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UserProfileV1' } } } }
        '400': { $ref: '#/components/responses/BadRequestV1' }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
  /api/v1/users/me/avatar:
    post:
      summary: Upload my avatar
      description: Uploads a profile avatar image and returns its storage key.
      tags: [ users ]
      operationId: uploadMyAvatar
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        description: Request payload for the uploadMyAvatar.
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/AvatarUploadFormV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AvatarUploadResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
  /api/v1/users/{id}/grade:
    patch:
      summary: Update user grade
      description: Sets or clears a user's grade. An admin may edit any user; a Team Lead may edit only users who are members of a group they lead. Re-evaluates standing group roadmap assignments so the user is enrolled into any assignment their new grade now matches, without removing existing enrollments.
      tags: [ users ]
      operationId: updateUserGrade
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the updateUserGrade.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserGradeRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateUserGradeResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
...
  /api/v1/teams:
    get:
      summary: List teams
      description: Returns teams visible to the caller along with assignable users and permission flags.
      tags: [ teams ]
      operationId: listTeams
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - name: teamsPage
          in: query
          required: false
          description: Zero-based page index for visible teams.
          schema: { type: integer, format: int32, minimum: 0, default: 0 }
        - name: teamsSize
          in: query
          required: false
          description: Maximum number of teams returned in one response.
          schema: { type: integer, format: int32, minimum: 1, maximum: 100, default: 20 }
        - name: teamsQuery
          in: query
          required: false
          description: Case-insensitive search by team lead name or email.
          schema: { type: string, maxLength: 200 }
        - name: usersPage
          in: query
          required: false
          description: Zero-based page index for assignable users.
          schema: { type: integer, format: int32, minimum: 0, default: 0 }
        - name: usersSize
          in: query
          required: false
          description: Maximum number of assignable users returned in one response.
          schema: { type: integer, format: int32, minimum: 1, maximum: 100, default: 20 }
        - name: usersQuery
          in: query
          required: false
          description: Case-insensitive search by user name or email.
          schema: { type: string, maxLength: 200 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TeamsResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
  /api/v1/teams/{leadId}/members:
    post:
      summary: Add team member
      description: Adds a member to the team led by the specified team lead.
      tags: [ teams ]
      operationId: addTeamMember
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: leadId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the addTeamMember.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddTeamMemberRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AddTeamMemberResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
    delete:
      summary: Remove team member
      description: Removes a member from a team using email or member reference in the body.
      tags: [ teams ]
      operationId: removeTeamMember
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: leadId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the removeTeamMember.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RemoveTeamMemberRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/teams/{leadId}/members/{memberId}:
    delete:
      summary: Remove team member by id
      description: Removes a member from a team using the member user id path parameter.
      tags: [ teams ]
      operationId: removeTeamMemberById
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: leadId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: memberId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
...
  /api/v1/grades/{id}:
    put:
      summary: Update grade
      description: Renames an existing grade. Requires grades.manage.
      tags: [ grades ]
      operationId: updateGrade
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the updateGrade.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateGradeRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GradeResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
    delete:
      summary: Deactivate grade
      description: Deactivates a grade so it no longer appears in active-grade lists. Existing users keep their current grade reference. Requires grades.manage.
      tags: [ grades ]
      operationId: deactivateGrade
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedV1:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
    ForbiddenV1:
      description: Authenticated but lacking permission
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
    NotFoundV1:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
    BadRequestV1:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
...
  schemas:
    PageInfoV1:
      description: Pagination metadata attached to a paged list response.
      type: object
      additionalProperties: false
      required: [ page, size, totalElements, totalPages, hasNext, hasPrevious ]
      properties:
        page:
          description: Zero-based index of the page returned in this response.
          type: integer
          format: int32
        size:
          description: Maximum number of items requested per page.
          type: integer
          format: int32
        totalElements:
          description: Total number of items matching the query across all pages.
          type: integer
          format: int64
        totalPages:
          description: Total number of pages available for the query.
          type: integer
          format: int32
        hasNext:
          description: Whether a page after this one exists.
          type: boolean
        hasPrevious:
          description: Whether a page before this one exists.
          type: boolean
...
    UserRoleCodeV1:
      description: User role code.
      type: string
      enum: [ admin, teamlead, member ]
...
    ApiErrorV1:
      description: Standard API error payload.
      type: object
      additionalProperties: false
      required: [ code, message, timestamp ]
      properties:
        code:
          description: Machine-readable code value.
          type: string
        message:
          description: Human-readable message text.
          type: string
        timestamp:
          description: Timestamp for the event or error.
          type: string
          format: date-time
        correlationId:
          description: Correlation identifier for tracing.
          type: string
        parameters:
          description: Parameters for Api Error.
          type: array
          items: { $ref: '#/components/schemas/ValidationParameterV1' }
    ValidationParameterV1:
      description: Single validation parameter attached to an API error.
      type: object
      additionalProperties: false
      required: [ code, value ]
      properties:
        code:
          description: Machine-readable code value.
          type: string
        value:
          description: Parameter value.
          type: string
...
    UserSummaryV1:
      description: Compact user profile shown in team and assignment views.
      type: object
      additionalProperties: false
      required: [ id, email, role ]
      properties:
        id:
          description: Unique identifier.
          type: integer
          format: int64
        name:
          description: Display name.
          type: string
        email:
          description: Email address.
          type: string
          format: email
        role:
          description: Role code for the user or member.
          allOf:
            - $ref: '#/components/schemas/UserRoleCodeV1'
        position:
          description: Job position or title.
          type: string
        avatarStorageKey:
          description: Storage key for the user's avatar image.
          type: string
        avatarColor:
          description: Fallback avatar color when no image is set.
          type: string
        gradeId:
          description: Current grade identifier, when a grade has been assigned.
          type: integer
          format: int64
          nullable: true
        gradeCode:
          description: Current grade's stable code, when a grade has been assigned.
          type: string
          nullable: true
        gradeName:
          description: Current grade's display name, when a grade has been assigned.
          type: string
          nullable: true
    UserProfileV1:
      description: User profile wrapper containing the current user summary.
      type: object
      additionalProperties: false
      required: [ user ]
      properties:
        user:
          description: Nested user payload.
          allOf:
            - $ref: '#/components/schemas/UserSummaryV1'
    UpdateProfileRequestV1:
      description: Fields for updating the signed-in user's profile.
      type: object
      additionalProperties: false
      properties:
        name:
          description: Display name.
          type: string
        position:
          description: Job position or title.
          type: string
        avatarStorageKey:
          description: Storage key for the user's avatar image.
          type: string
        avatarColor:
          description: Fallback avatar color when no image is set.
          type: string
    AvatarUploadFormV1:
      description: Multipart form for uploading a profile avatar image.
      type: object
      additionalProperties: false
      required: [ file ]
      properties:
        file:
          description: Binary file content.
          type: string
          format: binary
...
    AvatarUploadResponseV1:
      description: Avatar Upload Response payload.
      type: object
      additionalProperties: false
      required: [ storageKey ]
      properties:
        storageKey:
          description: Object storage key for an uploaded file.
          type: string
    OkResponseV1:
      description: Ok Response payload.
      type: object
      additionalProperties: false
      required: [ ok ]
      properties:
        ok:
          description: Indicates whether the operation succeeded.
          type: boolean
...
    PermissionMapV1:
      description: Permission Map payload.
      type: object
      additionalProperties: { type: boolean }
...
    TeamV1:
      description: Team payload.
      type: object
      additionalProperties: false
      required: [ members ]
      properties:
        lead:
          description: Lead for Team.
          allOf:
            - { $ref: '#/components/schemas/UserSummaryV1' }
          nullable: true
        members:
          description: Members for Team.
          type: array
          items: { $ref: '#/components/schemas/UserSummaryV1' }
...
    TeamsResponseV1:
      description: Teams Response payload.
      type: object
      additionalProperties: false
      required: [ teams, teamsPage, users, usersPage, permissions ]
      properties:
        teams:
          description: Teams for Teams Response.
          type: array
          items: { $ref: '#/components/schemas/TeamV1' }
        teamsPage:
          description: Pagination metadata for the teams collection.
          allOf:
            - $ref: '#/components/schemas/PageInfoV1'
        users:
          description: Users for Teams Response.
          type: array
          items: { $ref: '#/components/schemas/UserSummaryV1' }
        usersPage:
          description: Pagination metadata for the users collection.
          allOf:
            - $ref: '#/components/schemas/PageInfoV1'
        permissions:
          description: Permission flags keyed by permission code.
          allOf:
            - $ref: '#/components/schemas/PermissionMapV1'
    AddTeamMemberRequestV1:
      description: Add Team Member Request payload.
      type: object
      additionalProperties: false
      properties:
        memberId:
          description: Team member user identifier.
          type: integer
          format: int64
        member:
          description: Member lookup reference such as email.
          type: string
        email:
          description: Email address.
          type: string
    AddTeamMemberResponseV1:
      description: Add Team Member Response payload.
      type: object
      additionalProperties: false
      required: [ member ]
      properties:
        member:
          description: Member lookup reference such as email.
          allOf:
            - $ref: '#/components/schemas/UserSummaryV1'
    RemoveTeamMemberRequestV1:
      description: Remove Team Member Request payload.
      type: object
      additionalProperties: false
      required: [ memberId ]
      properties:
        memberId:
          description: Team member user identifier.
          type: integer
          format: int64
...
    UpdateUserGradeRequestV1:
      description: Update User Grade Request payload.
      type: object
      additionalProperties: false
      properties:
        gradeId:
          description: New grade id, or omitted/null to clear the grade.
          type: integer
          format: int64
          nullable: true
    UpdateUserGradeResponseV1:
      description: Update User Grade Response payload.
      type: object
      additionalProperties: false
      required: [ user ]
      properties:
        user:
          description: The updated user's profile.
          allOf:
            - $ref: '#/components/schemas/UserSummaryV1'
    GradeV1:
      description: A configurable grade dictionary entry.
      type: object
      additionalProperties: false
      required: [ id, code, name, displayOrder, isActive ]
      properties:
        id:
          description: Unique identifier.
          type: integer
          format: int64
        code:
          description: Stable internal code, derived once from the name at creation time.
          type: string
        name:
          description: Display name.
          type: string
        displayOrder:
          description: Sort order for listing.
          type: integer
          format: int32
        isActive:
          description: False when the grade has been deactivated.
          type: boolean
...
    UpdateGradeRequestV1:
      description: Update Grade Request payload.
      type: object
      additionalProperties: false
      required: [ name ]
      properties:
        name:
          description: New display name. 1-100 characters.
          type: string
          minLength: 1
          maxLength: 100
    GradeResponseV1:
      description: Grade Response payload.
      type: object
      additionalProperties: false
      required: [ grade ]
      properties:
        grade:
          description: The created or updated grade.
          allOf:
            - $ref: '#/components/schemas/GradeV1'
...